+++ src/day02.cpp
@@ -14,10 +14,10 @@ export module day02;
 
 export namespace day02 {
 
-std::tuple<int, int> solve(std::stringstream &ss);
-std::vector<std::vector<int>> parse(std::stringstream &ss);
-int part1(std::vector<std::vector<int>> const &vvs);
-int part2(std::vector<std::vector<int>> const &vvs);
+std::tuple<long, long> solve(std::stringstream &ss);
+std::vector<std::vector<long>> parse(std::stringstream &ss);
+long part1(std::vector<std::vector<long>> const &vvs);
+long part2(std::vector<std::vector<long>> const &vvs);
 
 }  // namespace day02
 
@@ -26,14 +26,14 @@ module :private;
 
 namespace day02 {
 
-std::tuple<int, int> solve(std::stringstream &ss) {
+std::tuple<long, long> solve(std::stringstream &ss) {
     auto input_data = parse(ss);
 
     return {part1(input_data), part2(input_data)};
 }
 
-std::vector<std::vector<int>> parse(std::stringstream &ss) {
-    std::vector<std::vector<int>> result;
+std::vector<std::vector<long>> parse(std::stringstream &ss) {
+    std::vector<std::vector<long>> result;
 
     for (std::string line; std::getline(ss, line);) {
         auto nums = util::getNumbers(line);
@@ -44,7 +44,7 @@ std::vector<std::vector<int>> parse(std::stringstream &ss) {
     return result;
 }
 
-int part1(std::vector<std::vector<int>> const &vvs) {
+long part1(std::vector<std::vector<long>> const &vvs) {
     int result{0};
 
     // Note: `vs` is a list of int in descending order.
@@ -56,7 +56,7 @@ int part1(std::vector<std::vector<int>> const &vvs) {
 }
 
 // Note: `vs` is a list of int in descending order.
-int findQuotient(std::vector<int> const &vs) {
+long findQuotient(std::vector<long> const &vs) {
     auto n_elm{vs.size()};
     for (std::size_t i = 0; i < n_elm - 1; ++i) {
         for (std::size_t j = i + 1; j < n_elm; ++j) {
@@ -69,8 +69,8 @@ int findQuotient(std::vector<int> const &vs) {
     return 0;
 }
 
-int part2(std::vector<std::vector<int>> const &vvs) {
-    int result{0};
+long part2(std::vector<std::vector<long>> const &vvs) {
+    long result{0};
 
     for (auto const &vs : vvs) {
         result += findQuotient(vs);
